(#11422) libiconv: add version 1.17
