+++ recipes/libiconv/all/conandata.yml
@@ -1,4 +1,7 @@
 sources:
+  "1.17":
+    url: "https://ftp.gnu.org/gnu/libiconv/libiconv-1.17.tar.gz"
+    sha256: "8f74213b56238c85a50a5329f77e06198771e70dd9a739779f4c02f65d971313"
   "1.16":
     url: "https://ftp.gnu.org/gnu/libiconv/libiconv-1.16.tar.gz"
     sha256: "e6a1b1b589654277ee790cce3734f07876ac4ccfaecbee8afa0b649cf529cc04"
@@ -6,6 +9,9 @@ sources:
     url: "https://ftp.gnu.org/gnu/libiconv/libiconv-1.15.tar.gz"
     sha256: "ccf536620a45458d26ba83887a983b96827001e92a13847b45e4925cc8913178"
 patches:
+  "1.17":
+    - patch_file: "patches/0001-libcharset-fix-linkage.patch"
+      base_path: "source_subfolder"
   "1.16":
     - patch_file: "patches/0001-libcharset-fix-linkage.patch"
       base_path: "source_subfolder"
