+++ recipes/libiconv/config.yml
@@ -1,4 +1,6 @@
 versions:
+  "1.17":
+    folder: all
   "1.16":
     folder: all
   "1.15":
